requirements.txt
absl-py==2.1.0
alembic==1.14.1
aniso8601==9.0.1
annotated-types==0.7.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
asyncpg==0.30.0
attrs==25.3.0
Automat==24.8.1
babel==2.17.0
beautifulsoup4==4.13.3
black==25.1.0
bleach==6.2.0
blinker==1.9.0
blis==1.2.0
bokeh==3.7.0
cachetools==5.5.2
catalogue==2.0.10
catboost==1.2.7
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
confection==0.1.5
constantly==23.10.4
contourpy==1.3.1
cryptography==44.0.2
cssselect==1.3.0
cycler==0.12.1
cymem==2.0.11
dash==3.0.0
dask==2025.2.0
databricks-sdk==0.46.0
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
docker==7.1.0
executing==2.2.0
fastapi==0.103.2
fastjsonschema==2.21.1
filelock==3.18.0
Flask==3.0.3
flatbuffers==25.2.10
fonttools==4.56.0
fqdn==1.5.1
fsspec==2025.3.0
future==1.0.0
gast==0.4.0
gensim==4.3.3
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphene==3.3
graphql-core==3.2.6
graphql-relay==3.2.0
graphviz==0.20.3
greenlet==3.1.1
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
hyperlink==21.0.0
hyperopt==0.2.7
idna==3.10
importlib_metadata==8.6.1
incremental==24.7.2
ipykernel==6.29.5
ipython==8.23.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
isoduration==20.11.0
itemadapter==0.11.0
itemloaders==1.3.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==2.13.1
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
libclang==18.1.1
lightgbm==4.6.0
locket==1.0.0
lxml==5.3.1
Mako==1.3.9
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
ml_dtypes==0.5.1
mlflow==2.21.0
mlflow-skinny==2.21.0
mpmath==1.3.0
murmurhash==1.0.12
mypy_extensions==1.1.0
namex==0.0.8
narwhals==1.31.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.3.3
notebook_shim==0.2.4
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.11.0.86
opentelemetry-api==1.31.0
opentelemetry-sdk==1.31.0
opentelemetry-semantic-conventions==0.52b0
opt_einsum==3.4.0
optree==0.14.1
optuna==4.2.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parsel==1.10.0
parso==0.8.4
partd==1.4.2
pathspec==0.12.1
patsy==1.0.1
pillow==11.1.0
platformdirs==4.3.7
plotly==6.0.1
polars==1.25.2
preshed==3.0.9
prometheus_client==0.21.1
prompt_toolkit==3.0.50
Protego==0.4.0
protobuf==4.25.6
psutil==7.0.0
pure_eval==0.2.3
py4j==0.10.9.9
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyDispatcher==2.0.7
Pygments==2.19.1
pyOpenSSL==24.3.0
pyparsing==3.2.1
pypdf==5.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
pytz==2025.1
pywin32==310
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.3.0
queuelib==1.7.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.10.1
Scrapy==2.12.0
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==3.4.1
service-identity==24.2.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.24
sqlparse==0.5.3
srsly==2.5.1
stack-data==0.6.3
starlette==0.27.0
statsmodels==0.14.4
stringcase==1.2.0
sympy==1.13.3
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
terminado==0.18.1
tf-keras==2.15.0
thinc==8.3.4
threadpoolctl==3.6.0
tinycss2==1.4.0
tldextract==5.1.3
tokenizers==0.21.1
toolz==1.0.0
torch==2.1.2
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
Twisted==24.11.0
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.1
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.0
w3lib==2.3.1
waitress==3.0.2
wasabi==1.1.3
watchfiles==1.1.0
watchgod==0.8.2
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wordcloud==1.9.4
wrapt==1.17.2
xgboost==3.0.0
xyzservices==2025.1.0
zipp==3.21.0
zope.interface==7.2
